airflow_base_gcp/credentials/requirements.txt
python-dotenv==1.0.0
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-google==10.14.0